pyproject.toml
[project]
name = "fastapi-project"
version = "0.1.0"
description = ""
authors = [
    {name = "Hammad",email = "hammadmubeen430@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.115.8,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "sqlalchemy (>=2.0.38,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "alembic (>=1.14.1,<2.0.0)",
    "pydantic (>=2.0,<3.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "python-jose[cryptography] (>=3.3.0,<4.0.0)",
    "celery (>=5.4.0,<6.0.0)",
    "redis (>=5.2.1,<6.0.0)",
    "pytest (>=8.3.4,<9.0.0)",
    "pytest-asyncio (>=0.25.3,<0.26.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "bump-pydantic (>=0.8.0,<0.9.0)",
    "email-validator (>=2.2.0,<3.0.0)",
    "anyio (>=4.8.0,<5.0.0)",
    "pytest-trio (>=0.8.0,<0.9.0)",
    "sentry-sdk[fastapi] (>=2.21.0,<3.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "bcrypt (==4.0.1)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest-asyncio = "^0.25.3"
pre-commit = "^4.1.0"

